+++ src/collective/cover/tiles/carousel.py
@@ -32,6 +32,7 @@ class ICarouselTile(IListTile):
 
 
 class CarouselTile(ListTile):
+
     implements(ICarouselTile)
 
     index = ViewPageTemplateFile('templates/carousel.pt')